+++ backend/tests/unit/page-role-links.test.ts
@@ -6,7 +6,7 @@
  *
  * @see backend/src/modules/seo/types/page-role.types.ts - ALLOWED_LINKS
  */
-import { PageRoleValidatorService } from '../../src/modules/seo/services/page-role-validator.service';
+import { PageRoleValidatorService } from '../../src/modules/seo/validation/page-role-validator.service';
 import {
   PageRole,
   isLinkAllowed,
